[ADD]ADD FileUser And FileMD5 Table

+++ disk-core/src/main/java/xyz/leyuna/disk/model/co/UserFileInfoCO.java
@@ -17,7 +17,7 @@ public class UserFileInfoCO {
     /**
      * 用户内存总量
      */
-    private Double fileTotal;
+    private Long fileTotal;
 
     /**
      * 用户文件
